3맵 오두막 바깥 대사 완료 --> player 수정 필요
