+++ Assets/JeongJH/Script/Dialog.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 96214c799ef098f4bb6fe68138d90e51
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
