+++ Assets/JeongJH/Script/Dialog/ch2Dialog.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 07d300c2bffa8264db9d8a4f4c6b3a22
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
