+++ Assets/JeongJH/Script/Dialog/ch3_1Dialog.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ae309a7e99b6085408640f01e95c6859
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
