+++ Assets/JeongJH/Script/Dialog/ch3_1ObjectDialog.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0a734788642e6e849b069018f3d3d2d3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
