--- Assets/JeongJH/ch1Dialog.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 03f98898e9f8f3f4abffa33fcfefab29
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
