--- Assets/JeongJH/ch2Dialog.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 07d300c2bffa8264db9d8a4f4c6b3a22
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
